lambda_functions/s3_to_rds_lambda/src/requirements.txt
psycopg2-binary<=2.9.9